Add files via upload
I couldn't find the names of the graphics files that they need to point to, but I've typed in just the 5 core enhancement names and descriptions for the vistani knife fighter tree for a quick start. Please take a look and let me know if this will help.  Also I'm doing the editing on a mac while looking at ddo on PC.  I used visual code so I don't think line endings are impacted but will verify line endings after upload.